requirements.txt
Django==1.10.5
django-bootstrap3==8.1.0
djangorestframework==3.4.7
django-widget-tweaks==1.4.1
pillow